updated the Jenkins version, added a memory limit to the JVM
It shall now always fetch the latest version Jenkins.
The previous version was a huge memory hog, updating it fixed it. I
also added a hard limit on the JVM, is runs fine for me and should help
it running well on sandbox mode.

Updated the readme to the right repo.